Post and Comment Ratings System
In order to allow visitors to rate posts and comments, links were added
to the article show view, and the comment partial to allow users to
upvote and downvote a comment.

Additionally the rating of the posts and comments can be seen on the
same pages with a breakdown of the current rating into upvotes and
downvotes.